moved more ts files to vue
